requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
firebase-admin==6.3.0
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
gcloud==0.17.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h5py==3.10.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.7
numpy>=1.26.2
oauth2client==3.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.10.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-jwt==2.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
